mira configuración de log a un archivo propio
