+++ core/mpi/controller/mpiUpdater.ts
@@ -1,7 +1,7 @@
 import { userScheduler } from '../../../config.private';
 import * as controller from './paciente';
 import { paciente, pacienteMpi } from '../schemas/paciente';
-import { logKeys } from '../../../config';
+import { logKeys } from '../../../config-log';
 import * as servicioAnses from './../../../utils/servicioAnses';
 import { log } from '@andes/log';
